Move auto-revoke policy to hibernation policy (2/2)
Rename several auto-revoke policy methods to hibernation. There are a
few places where a more thoughtful approach will be needed (i.e. an SDK
target check most likely) but these are left as is and marked with TODOs
for now. They will be addressed in further CLs.

Also removes a few unused variables and methods

Bug: 175830282
Test: atest AutoRevokeTest
Test: adb shell cmd jobscheduler run -u 0 -f com.google.android.permissioncontroller 2

Change-Id: Ic8a5c5ae42f775546b98ef660e7cc1a8fb6b1c03