Check if admins may grant sensors permissions
Check whether the device policy client (DPC) may grant sensors-related
permissions.

When a permission is explicitly granted, the DevicePolicyManagerService
will pass a flag indicating if the admin may control grants for
sensors-related permissions.

When the grant may come via policy, the GrantPermissionsViewModel will
call back the DPMS to find out if the admin has this level of control.

If the admin is trying to grant a permission it may not control, treat
it as the "default" state (which will lead to a runtime permission
prompt being shown when the app runs and requests the permission).

Manual testing:
* Install TestDPC
* Create a work profile using TestDPC.
* Get the BasicLocation app by checking out
  https://github.com/android/location-samples and building it from there.
* Install the app onto the device but do not start it.
* In TestDPC, Find "Manage app permissions", choose "Basic Location Sample"
  from the drop-down menu.
* Toggle each of the "ACCESS_COARSE_LOCATION" and
 "ACCESS_BACKGROUND_LOCATION" to "Allow".
* Observe that no notification appears.
* Start the BasicLocation app and observe the runtime permission prompt
  shows up.

Bug: 158735247
Test: Manual, see above
Test: atest com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testGrantOfSensorsRelatedPermissions
Test: atest com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testDenyOfSensorsRelatedPermissions
Test: atest com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testSensorsRelatedPermissionsNotGrantedViaPolicy
Test: atest com.android.cts.devicepolicy.MixedDeviceOwnerTest#testGrantOfSensorsRelatedPermissions
Test: atest com.android.cts.devicepolicy.MixedDeviceOwnerTest#testSensorsRelatedPermissionsNotGrantedViaPolicy
Change-Id: Ibb2c0f3ae2bf2162680231d86b6fa9105099af79