Set GRANTED_BY_ROLE flag for app op permissions granted by role.
Previously we didn't utilize this flag for app op permissions because we
were considering whether install permissions and runtime permissions
should be separated. However since the introduction of role protected
permissions, which are install permissions, we are already using
permission flags on install permissions. And the unification of install
permission and runtime permission storage in S also made this natural.
So let's utilize the GRANTED_BY_ROLE flag for app op permissions now.

Fixes: 180431636
Test: presubmit
Change-Id: I418e02f83950e101573d6d18962f287d242ce983