Add minSdkVersion attribute for role.
This is needed as we come to the second platform release since role
was first modularized in R, because there's no reason to retroactively
introduce the new S roles back to R.

This change adds a minSdkVersion attribute which defaults to BASE (1)
and sets it to S (31) for all the new S roles. Related documentation
is also updated to reflect this change.

Fixes: 180761423
Test: presubmit
Change-Id: Ic34d9904365b028300e190b1cf38bcdc1b97e007