Implement collapsing toolbar design on S+ for role UI.
The UI code for PermissionController is a bit complex because we need
to support different form factors, as well as different platform
versions now.

In the case of role UI, the code was structured in the following way:

1. Activities are just stubs, and loads form factor specific fragments
to provide the actual UI content.

2. Although we want to share preference business logic among different
form factors, the fact that Auto needs a CarPreferenceFragment base
class while normally we use vanilla AndroidX PreferenceFragmentCompat
for phone makes this complex. The solution was to create a child
fragment that requires its parent to be a PreferenceFragment and
provides certain callbacks, then different preference fragments for
different form factors will add this child fragment and easily share
most of the implementation.

3. For phone UI, we were using the system action bar so
the base class SettingsFragment can simply extend from
PreferenceFragmentCompat and be added directly to the
activity. However since we are adding additional non-preference UI
wrapping the preference UI, i.e. CoordinatorLayout and
CollapsingToolbarLayout etc, the proper way to do this is to introduce
an additional parent fragment acting between the activity and the
preference fragment. So take default app list for example, the
hierarchy is now:

    - DefaultAppListActivity (stub activity)
        - HandheldDefaultAppListFragment (toolbar, loading/empty)
            - HandheldDefaultAppListPreferenceFragment (preference)
                - DefaultAppListChildFragment (business logic)

Special app access code, which is unused for now, are updated in the
same way in case we want to enable it in the future.

Appropriate changes are also made to theming in a backwards compatible
way.

This change hides the global toolbar actions on S+ as well.

Bug: 177638625
Test: manual
Change-Id: I001d048b851f7037a8455fe00954501826fe3f7a