Use component usage metric when hibernation enabled
For S devices, we want to use the new component usage metric in
addition to last time visible so we can include implicit usage
as well. This is the case for both auto-revoke and hibernation. For
pre-S devices, we want to continue to use last time visible as
before.

For now, this is hidden behind the hibernation feature flag but should
be updated when the current SDK moves up.

This is code covered by the existing AppHibernationIntegrationTest

Bug: 175830282
Test: manual, ran job after a service binding, app not hibernated
Change-Id: I122634453e8ff1de9437158ee00bdcdb8830c6d5