Fall back to system default browser when the current default is no longer valid.
To adapt to the new UX requirements in S. The change in behavior is
guarded by an SDK check.

Note that the system default browser here for the purpose of fallback
can be set either explicitly via config_defaultBrowser, or implicitly
by being the one and only one browser on the system image.

Bug: 183412485
Test: manual
Test: Install 2 3p browsers, select 1 as default and uninstall, then
      the system browser becomes the default browser
Test: Install 2 3p browsers, select the system browser as default and
      disable, then no browser is default. Enable the system browser
      again then the system browser becomes the default browser.
Change-Id: I672b73356659b5162b279f3abf418001007e371c