Interpret 31+ for minSdkVersion in roles.xml as S+.
There is currently no way to tell if a build is master or sc-dev for now, but we are starting to work on T-only roles on master. As a workaround, we can use an integer larger than 31 for T, but still only check if the build is at least S for now, until we have the T codename on master.

Bug: 158878560
Change-Id: I82050ea923f452340203a37f50f0a1f935166bf1
Test: presubmit