Hibernation only targets S apps
Limit hibernation to only apps that target S. Since this now introduces
the possibility of apps being auto-revoked but not hibernated, modify
the unused apps live data to include the union of auto-revoked apps and
hibernated apps.

This code is covered by a new test in AppHibernationIntegrationTest
introduced in this topic.

Bug: 183141572
Test: atest AppHibernationIntegrationTest
Test: manual

Change-Id: Ibc8ec5c339111ff90befeb3ea64f69088195ff15