Allow apps with dynamic libs to hold roles
Apps exposing dynamic libs should be able to hold roles as
such an app can have other functionality in additon to being
a lib. For static shared libs (cannot have anything by the
lib in the APK) and for built-in libs (these are jar files)
we don't support being a role holder.

While at this also make sure the AR role holder handles the
new ACTIVITY_RECOGNIZER intent in a service as this is used
to specify AR related exampt tags.

bug: 179069158

Test: system roles work with apps that also have a lib

Change-Id: Ib6ce9bfd258845720a22ee68b626f4b0332f7168