Don't auto-revoke implicit permissions
Recently, we made Bluetooth a split-permission. The result is that:

- targetSdk>=S apps must explicitly request the Bluetooth runtime
  permission, but
- targetSdk<S apps will be implicitly granted the Bluetooth runtime
  permission, provided they request the Bluetooth install-time
  permission, and
- we now only "check" the Bluetooth runtime permission when Bluetooth
  functionality is needed

We then encountered an issue where the auto-revoke feature, which was
designed to automatically revoke runtime permissions which haven't been
used in a while, was auto-revoking the Bluetooth runtime permission.

For apps with targetSdk<S, this is especially a problem, because it
means legacy applications that use Bluetooth may have Bluetooth revoked
and, since they have never had a chance to code around this possibility,
these apps will most likely crash after an auto-revoke occurs.

The solution is to modify how we auto-revoke: if a permission was
implicitly granted (which is true if-and-only-if the permission has the
"REVOKE_WHEN_REQUESTED" flag) then we will not auto-revoke it.

Bug: 184962595
Test: atest AutoRevokeTest
Change-Id: Iee949c8c7d1b106274c3279a1dc41105a603daaa