Add flag for hibernation to target pre-S apps for teamfood
Add a device config flag for teamfood study that allows us to
override the default hibernation exemption setting for pre-S apps. This
allows hibernation to hibernate all apps by default on the system. It
also makes it so that full hibernation takes effect on pre-S apps
instead of just doing auto-revoke.

Bug: 186261550
Test: manual
Change-Id: I5b34ea2e17e6be69096d25e81e654bba90731144