Add minSdkVersion for permissions in role.
For permissions new in S, they are non-existent on R and trying to
grant them with result in an IllegalArgumentException from
PermissionManagerService. So add a minSdkVersion attribute to make
such new permissions only managed when available.

Also skip parsing test for roles that are only available on newer
SDK versions, which may contain references to newer things including
newer permissions.

Bug: 182960931
Fixes: 186724565
Test: atest PermissionControllerMockingTest
Change-Id: I5c90db23d2474833731de22521658794b1481819