Remove duplicate AndroidX dependencies from test APK.
The test is an instrumentation test so its dex will be loaded into the
target process during instrumentation, and the test dex will have a
higher precedence than the target dex in the dexPathList of the
PathClassLoader.

However, when the target code is loading a resource in a library it
included, it will be loading the resources from a context of the
target app. And in this case, if the test dex contains a duplicate R
class for the library, that test R class will take precedence and
won't be able to load the resource from the target app's resources,
resulting in a crash.

In this case, com.google.android.material.internal.R is being
duplicated in both the target app and the test app, so when
ThemeEnforcement tries to resolve the theme attribute of
R.attr.colorPrimary, it is resolving the test app R value against the
target app (context's) resources, and caused the test failure.

The proper solution is to never include duplicate AAR libraries in
test app, and we need to exclude the transitive Material Components
dependency introduced by androidx.test.espresso.contrib. As currently
there is no way to do the exclusion in Soong, this can be done by
depending on the androidx.test.espresso.contrib-nodeps target instead.

Bug: 183459058
Fixes: 188004185
Test: atest PermissionControllerInProcessTests
Change-Id: I85e67fa618f10abc8a319b324b11914992a7e0d2