Fix uninstall runable not updating correctly
The remove runnable was not updating when a new view holder was bound
for an unused app preference in the unused apps page. This caused the
stale runnable to stay on the button and caused incorrect apps to be
uninstalled when the user tried to uninstall an app from the page.

Bug: 188376857
Test: manual
Change-Id: I097474f9113100c6b3a2987d40a7f976796463bd