Use application context to create cached ServiceConnectors.
This used to be using ActivityThread.currentApplication(), but was
changed to the context passed in in ag/13387619 in S and caused the
regression. We cannot use an activity context here if we are to cache
the connectors, because removed activities can't bind to services, so
use Context.getApplicationContext() to get back to the R behavior.

Fixes: 190449945
Test: manually with the reproduction steps
Change-Id: I73702f932816977104b17aec52cec628036218c6