Show string when no apps in unused apps
Show a string when there are no unused apps instead of just
having an empty list with the footer.

This wasn't happening before because the footer preference counted as an
item in the recycler view adapter, and as long as it was visible, the
empty view in PermissionsFrameFragment would not show up. By updating
the visibility of this footer, the string then shows correctly.

Bug: 190829147
Test: run hibernation job, re-grant all revoked permissions, see empty
string

Change-Id: I183f578094950f5c86e625117f6afa3188acf5ff