Use SdkLevel instead of BuildCompat in Permission.
SdkLevel is the recommended class to check for pre-release platform in
Mainline, and is also wider used in our current code. BuildCompat does a
similar thing but due to the complexity in SDK finalization and
AndroidX, its logic can go out of sync for a period of time until a new
prebuilt is imported, while SdkLevel can be fixed directly within AOSP.
So we should use SdkLevel instead of BuildCompat wherever possible in
our code.

Bug: 191114884
Test: presubmit
Change-Id: I82465cab7fba80709cded09a69af73e2d1eac86b