Add hibernation toggle string to "App Permissions"
The string used to toggle hibernation exemption for the app needs to
have the same text in "App info" and "App permissions". As the string
for "App info" is in Settings and isn't accessible to
PermissionController, we add the string to the PermissionController code
base.

This just introduces the string so that localization can be done sooner
and does not functionally change anything:

Bug: 191901459
Test: builds
Change-Id: I091ca2e649827f1a955fdd37c1fea6aa3658ead2