Make FAB a direct child of CoordinatorLayout to fix its position.
Inflating the FAB inside the fragment view, which is under the
scrolling view child of the CoordinatorLayout in
CollapsingToolbarBaseFragment, makes it impossible to position the FAB
as sticking to the page bottom. The official way of adding a FAB is to
add it as a direct child of the CoordinatorLayout, so we should
follow that best practice to solve our problem.

https://material.io/components/buttons-floating-action-button/android#extended-fabs

Fixes: 189423430
Fixes: 191508456
Test: manually check the FAB position, and test screen rotations.
Change-Id: I683855100ea42bc21e3fe1cd2baf1b89536898fb