Enable permission to perform its own hidden API processing
Previously, the hidden API encoding of the permission boot dex jars,
i.e. those dex jars that permission contributes to the bootclasspath
were done as part of the monolithic hidden API processing. This change
causes the encoding to be done by the permission's
bootclasspath_fragment.

This change involves the following:
* Addition of the fragments property to the permission's
  bootclasspath_fragment module to list all the other
  bootclasspath_fragment modules on which this depends.
* Addition of the additional_stubs property to add stubs for APIs that
  are not provided by another bootclasspath_fragment.

The build automatically checks that the hidden API flags which are
computed by permission and encoded into its boot dex jars match those
that are generated by the monolithic processing so this is guaranteed
to be safe.

Bug: 179354495
Test: m com.android.permission
      - ensure that the generated APEX is byte-for-byte identical
        before and after these changes.
      m out/soong/hiddenapi/hiddenapi-flags.csv
      - make sure that they are not changed by this.
Change-Id: Idcbdd331d5b5e897338bba46485cb56dab20dd4b