Hide focus until list has been loaded
This change addresses the rotary requirement to focus on the first list
item in a new screen. Since the content is loaded asynchronously we need
to add this logic to let the view know where to add focus once the
content has loaded. The Chassis utils provide a clean, easy way to
accomplish this.

Bug: 191303784
Test: Manual testing - navigate to Settings > Apps > Unused apps using
rotary
Test: m GooglePermissionController && adb install $OUT/system/priv-app/GooglePermissionController/GooglePermissionController.apk

Change-Id: I6c8d576bcc4ce122fb6a5081ec257e5491ccedad