Fix loading spinner not showing up
Loading spinner wasn't showing up because unused apps were preloaded,
and there is no logic to show anything on screen in this case.
Logically, we'd probably want to show the actual unused apps list if
unused apps are loaded, but there's actually an additional step to turn
unused apps into unused app categories which is its own live data.

So we change the check to see if unused apps categories live data is
loaded and show the view immediately if so. Otherwise, we always show
the loading view.

Bug: 193730546
Test: manual
Change-Id: If61647c9efa98ef45bb454ff62b83e5801c028d5