Add last access to Auto apps permission screen
Replaces the deprecated PermissionApps loader with the new
PermissionAppsViewModel. This brings the Auto fragment to closer parity
with the phone implementation. It also allows us to mirror the phone
implementation for "Last access" description.

To avoid code duplication and fragmentation, a number of a helper
methods have been moved into the ViewModel. They could go in a util
class or some other class, but this seems like a reasonable central
location that faciliatates code sharing.

The main difference between the phone and Auto implementation is that we
don't current show the scoped storage changes that were introduced in R.
We may integrate this feature in the future, but there isn't much media
usage on Auto, so it's not a high priority.

HSV: https://hsv.googleplex.com/4866387384205312

Bug: 191988264
Test: atest GtsPermissionControllerHostTestCases
Test: manual testing
Merged-In: I96efbc51d628ea12ce4b1f6de1ecce181e9fdc83
Change-Id: I96efbc51d628ea12ce4b1f6de1ecce181e9fdc83