Ensure attributes for using Toolbar are in Settings theme despite OEM overlays.
The original @style/Settings is made overlayable to OEMs for them to
customize the base theme in ag/7145826, however we do need a theme to
place required attributes such as the attributes for using Toolbar
instead of ActionBar. So introduce a new theme (following the common
naming scheme) called Theme.PermissionController.Settings and make it
inherit from the overlayable base theme Settings, and put the toolbar
attributes in it.

The same structural changes are applied to the other form factors,
except that they don't require the toolbar attributes.

Bug: 195715344
Bug: 195726506
Test: manual
Change-Id: I44d791d7527b39a1a3abeaa77936c2969e5f01c6