Don't start a new one-time permission session if not needed
Previously one-time permissions assumed that all permissions in the
group would be revoked when the permission expired, but in S it is
possible to grant coarse location permanently but fine as one-time.
The fix is to only start a one-time permission session if any of the
currently granted permissions have the ONE_TIME flag.

Test: Manual + `atest CtsPermission{,2,3}TestCases`
Fixes: 195961465
Change-Id: Ib3b26183ee5f9d591be7f9d45b542e974716805c