Shrink permission dialog window
Change the size of the dialog window to only be as large as the dialog
is. In order to still support the gravity parameter we use the gravity
value from grant_singleton which is set by the PermissionGrantSingleton
styleabl, we do this to stay backwards compatible with the previous
implementation, where this lineralayout was the thing that controlled
the dialogs veritcal alignment.

Bug: 200801402
Test: Manually verified that the window is just the size of the dialog
and that it is positioned at the bottom when bottom|center|top according
to the PermissionGrantSingleton styleable
Test: Verified that the permissiondialog scoll functionallity is intact
by using splitscreen to shrink the availble area

Change-Id: I78d8c78863bd6ff9c0b105595420e91b6b741754
Merged-In: I78d8c78863bd6ff9c0b105595420e91b6b741754