Add the core of the access checking service.
This change contains the core of the new unified and generic access
checking subsystem, which will be backing both the permission and the
app op API, as well as any future features like per-attribution
permissions.

The AccessUri file defines different types of URIs which represents
different entities in the system, e.g. a UID, a package, a permission
or an app op. Each different type of URI has its own scheme, and for
each combination of subject (e.g. UID) and object (e.g. permission), a
corresponding policy can be defined to maintain the corresponding
state and return access check decision. The state is maintained in
SystemState and UserStates, for system-level and user-level
states. (Some parts of) them are also persisted upon changes, and
restored during boot.

The util package contains the collection class typedefs and extension
methods, so that standard Kotlin collection utilities can be used with
a concrete collection type which improves performance by avoiding
invoke-interface calls. All classes using collections are expected to
import com.android.permission.access.util.* in order to make use of
them instead of the standard ones which are defined for interfaces.

Bug: 182523293
Test: build
Change-Id: Ie82dadb54333f7ddf20eb440d603fefbac2da3f0