Default overrideUserWhenGranting to false.
The previous default value of true made sense when most of the roles
were default apps, and the user action of setting a default app
implies that the permissions should always be given to make the
default app work. However since nowadays most of the roles are for
pregrants, these pregrant roles should never override user's explicit
denial, and the previous default apps are marked as so explicitly
instead.

In other words, this CL changes all the non-defaults app roles to never
override user set permissions, but keeps allowing default app roles to
do so.

Fixes: 202312327
Test: presubmit
Change-Id: I8b862eebf61acf4cbb1399c9869054a1414a2394