Add support for system-called GrantPermissionActivity
Let the GrantPermissionsActivity handle the
REQUEST_PERMISSIONS_FOR_OTHER intent action. This will be called by the
system (and suppressed for anyone else) to show a permission prompt for
another app. Also add the ability to merge multiple requests for the
same app into one prompt, which then sends results back to the other
GrantPermissionActivities

Bug: 194833441
Test: atest NotificationPermissionTest

Change-Id: I9285ac2be5c60622d907a49c5d21e09d361f7177