Persist permission decisions for Automotive
For Automotive, persist permission decisions made in the grant dialog or
app permission settings for review by the user.

We considered persisting the data in SQLite, but the core framework team
advised against it due to unnnecessary memory overhead and general
complications that come with SQLite. Instead the data is serialized and
deserialized from XML and written/read with an AtomicFile.

The persisted data has a few characteristics:
- Unique for (package_name, permission_group)
- Data should only be stored for 7 days (to be added in follow-up cl)
- All methods are asynchonous since they involve file I/O

Follow-up work will include UI surfacing the decisions, clearing of
data for uninstalled packages and decisions older than 7 days, and
sending a reminder notification post-drive.

Bug: 194240664
Test: atest RecentPermissionDecisionsStorageImplTest
Change-Id: I683b903f539df5d16939d30fd1a592214f9f2ebc