Remove circular dependency context from interface
The context used in these parent/child interfaces is supposed to be
context coming from the parent. Having the parent get its own context
via the child is confusing and potentially leads to errors by a
developer mistakenly passing the wrong context to the parent. Removing
context from the interface makes it more likely that the parent uses its
own context, as intended.

Bug: 205905291
Test: Manual testing
Change-Id: I25a3e22ef304a16f5bc4dc7e873e08927279d6d6