Fix test failure due to R import.
There's magic happening with library R class processing and it seems to
be somehow breaking when we try to access it in an instrumentation
(maybe the reference isn't properly patched/replaced). Since all the
other code is referencing the app R class and this reference to the
AndroidX R is the only one failing, just change it to reference the app
R as well.

Test: atest GooglePermissionControllerInProcessTests:ManageStandardPermissionsFragmentTest
Change-Id: I05ef2afd849e65ae68d7dfc07c69ed730f84a73c