[Angel] Add support for PendingIntents to SafetySourceData and address comments from ag/16182563
* Adds SafetySourceIssue.Action representing button actions to safety issue, and add PendingIntents for safety preferences.
* Rename
    SafetyPreferenceData -> SafetySourceStatus, SafetyIssueData ->
    SafetySourceIssue, SafetyPreferenceData.SeverityLevel ->  SafetySourceStatus.StatusLevel.I
chose to keep the levels separate as their javadocs/guidelines for
choosing a level will be different.
* I haven't folded SafetyPreferenceData into SafetySourceData as the
SafetyPreferenceData will be @Nullable. UX has confirmed we want this
behaviour for sources - sources can provide issues without the
preference, but in this case the warnings will not be allowed to affect
the overall status.
* Refine javadocs for API
* Delete aidl files other than SafetySourceData as that's the only class
to be used as arguments in IPC

Test: Will add CTS tests when API is no longer hidden.
Bug: 206752239

Change-Id: I868bb324f14e2f4262b8796bccdde34f6aa9ba68