Cleanup job to remove old permission decision data
To protect the user's privacy, we limit the amount of permission
decision data that we keep. Default cap is to keep data for the last 7
days.

If necessary, schedule the clean-up job when the storage db instance is
created. Run the job once per day to check for decisions that should be
removed.

Timing for keeping old data and check frequency is configurable through
newly introduced DeviceConfig values.

Bug: 194240664
Test: atest PermissionControllerMockingTests
Change-Id: I39203bbb39f7e4e10dff7650f9e459efc3bb0cb3