Permission decision review screen
Create a screen to show recent permission decisions with the
corresponding fragment, viewmodel, and livedata. The livedata is a
wrapper around the database access. The ViewModel filters out permission
decisions that no longer reflect the current state and also does a
second check that the package is still installed.

Will add UI tests for fragment in a follow-up cl.

Screenshot: https://hsv.googleplex.com/5375245504479232

Bug: 194240664
Test: atest RecentPermissionDecisionsLiveDataTest
Change-Id: I84842d278d63f402b4d237b36799c10e4be0b74b