Don't filter permission groups by showSystem
After feedback from UX, we're not going to filter the permission groups
that are shown based on whether they are only used by a system app.
Initially the idea was that we wouldn't show groups that are irrelevant,
just as we did with Permission Decluttering, but since the UI here
updates dynamically we decided that it's more confusing in this case.

Also clarify that we're initially not showing last 7 days at this time.

Bug: 193446466
Test: Manual testing
Change-Id: Ic68b1e2e70bbd5c220f51329c9fd283e45a2737c