Fix self-revocation for background permissions.
This includes a fix to actually revoke background permissions at the end
of a one-time-permission session, and a UI fix to correctly display a
one-time background permission as "Ask every time" on the "App
permissions" page.
This now makes it consistent with the settings page showing the status
of an individual permission group. Both the permission group status page
and the app permissions page now have the same behavior for one-time and
self-revoked permissions (until the revocation is triggered on kill).

Bug: 210387494
Test: atest android.permission.cts.RevokeOwnPermissionTest
Change-Id: I951d4ff4d528088c47d511b7a41f82ce8064688a