Add new groups, and enforce media supergroup
Create the following permission groups and permissions:

    READ_MEDIA_AURAL (group):
        READ_MEDIA_AUDIO (permission)
    READ_MEDIA_VISUAL (group):
        READ_MEDIA_VIDEO (permission)
        READ_MEDIA_IMAGE (permission)

Also, combine the following three groups...

- STORAGE
- READ_MEDIA_AURAL
- READ_MEDIA_VISUAL

...into one "supergroup" for apps with targetSdk<TIRAMISU. This means,
when a user attempts to allow or deny one any one of these groups, the
others will also be allowed or denied (but only for apps with
targetSdk<TIRAMISU).

The motivation for this change is to support the "Media Permission
Update for Android T" project.

Bug: 195055433
Test: atest CtsPermissionTestCases CtsPermission3TestCases
Change-Id: Ia52543ece70ae9d9e4d5afd854272b4d2b10eb98