Obsolete "manage all files" preference categories
The STORAGE group currently displays these categories:

- Allowed to manage all files
- Allowed access to media only
- Not allowed

After this change, we'll simply have:

- Allowed
- Not allowed

This means that this screen will no longer be able to toggle the
MANAGE_EXTERNAL_STORAGE permission; this will need to be done on the
separate "all files access" screen.

The motivation for this change is to support the "Media Permission
Update for Android T" project.

Bug: 195055433
Test: atest CtsPermissionTestCases CtsPermission3TestCases
Change-Id: I1abab9e1f449f87320ed2859dd05bc0761b85346