Merge grant dialogs for media permissions
We recently combined the STORAGE, READ_MEDIA_AURAL, and
READ_MEDIA_VISUAL permission groups to act as a "supergroup" (for
targetSdk<T apps, specifically) which means when the state of one is
changed, the others will as well.

However, when this supergroup is allowed or denied, it results in three
GrantPermissions dialog promps. This is unnecessarily verbose. So, this
change merges the three successive grant dialogs into one.

The motivation for this change is to support the "Media Permission
Update for Android T" project.

Bug: 195055433
Test: atest CtsPermissionTestCases CtsPermission3TestCases
Change-Id: I19d87d54bfa56b5b824dc2ec6b9616bdf52df61b