Add warning dialogs when changing legacy storage permissions
When a user clicks "allow" or "deny" for a storage group (which now
includes STORAGE, READ_MEDIA_AURAL, or READ_MEDIA_VISUAL), and the app
is legacy (targetSdk<T), warn them with a confirm dialog that informs
them that changing one group will actually change the other two.

The motivation for this change is to support the "Media Permission
Update for Android T" project.

Bug: 195055433
Test: atest CtsPermissionTestCases CtsPermission3TestCases
Change-Id: I63c70c9da5fa66a561f9c366679c345eb6a13c3a