Pass userId from SafetyCenterManager to SafetyCenterService for all APIs.
This puts us in line with go/android-api-guidelines#use-context-user
API-wise but leaves implementation for a later date. Some APIs may
support a multi-user scenario, others may throw an error when the user
ID in the API doesn't match the caller's identity or belongs to a
non-full user.

I also included a little bit of reformatting to make method calls and
definitions more readable.

Bug: 216794581
Test: m com.google.android.permission && adb install $OUT/system/apex/com.google.android.permission.apex && adb reboot && atest CtsSafetyCenterTestCases
Change-Id: Id9cc9ff4a62d4c49db14a3c8c1ca92144d532171