Add hyperlink footer to all files access
Because we've obsoleted the "all files access" preference categories,
we'll need to transition users who are used to this feature over to the
"all files access" screen, and we do that here by adding a hyperlink.

The motivation for this change is to support the "Media Permission
Update for Android T" project.

Bug: 195055433
Test: atest CtsPermissionTestCases CtsPermission3TestCases
Change-Id: I5c482f7136f1f752b9d4278ba4df53fdddc86c32