Change self-revocation to work by permission instead of group
Previously, the revokeOwnPermissionsOnKill API would revoke the entire
group for each input permission. We now revoke only the permissions
listed, along with their background counterparts if there is no
granted foreground permission left.

Includes a number of adjustements to the permission settings UI to
account for the new state combinations (having both one time and
while-in-use permissions in the same group).

Test: atest android.permission.cts.RevokeOwnPermissionTest
Bug: 210387494
Change-Id: Id04af89fecb95b4bc18458b1bb741fb05409f3f1