Add APIs for errors and inflight actions to SafetyCenter.
All actions will be executed by the safety center service which can keep track
of in-flight resolving actions (aka inline actions). Sources will be
able to notify the Safety Center about errors when handling resolving
actions. Without this API, actions could only fail by timing out.

While the SafetySourceError API only supports action errors now, it
could be extended with other error types in the future.

Also adds a new method to OnSafetyCenterDataChangedListener which will
notify the Safety Center UI about errors to display to the user. This is
intentionally generic to minimize business logic in the UI and allow error
messaging logic to live within the SafetyCenterService.

In order to identify actions to execute and that errored, an ID field is added to
SafetyCenterIssue.Action and SafetySourceIssue.Action.

Implementation of these APIs will be in a separate change.

Change-Id: Ife3de58a103eba4b23157b1b057a7b68e27d5e18
Bug: 218379298
Test: atest CtsSafetyCenterTestCases