Add READ_CLIPBOARD_IN_BACKGROUND to the SYSTEM_TEXT_INTELLIGENCE role
Android System Intelligence is allowed to read the clipboard in the
background. Previously this was relying on a special case in
ClipboardService, checking that the calling package is the provider of
ContentCaptureService. In order to remove this special case, we add the
permission READ_CLIPBOARD_IN_BACKGROUND to the role
android.app.role.SYSTEM_TEXT_INTELLIGENCE which is held by Android
System Intelligence.

Bug: 216314180
Test: Presubmit
Change-Id: Icba3763e66b6694d87792315904ed785abe049f6