Disable SafetyCenter APIs if flag is off.
Part of this work is already in tm-mainline-prod, but we need this in
tm-dev since we don't want the APIs to be usable until we turn the flag
on.

The UI and the newly added tests will have to be merged in
tm-mainline-prod.

Bug: 223378357
Test: atest CtsSafetyCenterTestCases
Merged-In: Ie76840706294e031d00cb74b2d080de8a70774f7
Change-Id: Ie76840706294e031d00cb74b2d080de8a70774f7