Do not check SafetyCenterActivity on cars.
We sent a change to disable SafetyCenter on cars here: ag/17259584.
However, we have some tests for devices that don't support SafetyCenter
as well; in which case we still want to skip the
SafetyCenterActivityTest here since it redirects to the Security page
when SafetyCenter is disabled; but the security page is not available on
cars.

Bug: 225094063
Test: atest CtsSafetyCenterTestCases
Change-Id: I2830818f9b188bf1f7835a034f18f89cebec4b2e